Improve documentation and change style such that checkstyle passes
