+++ src/main/java/yoyo/Yoyo.java
@@ -7,6 +7,9 @@ import yoyo.core.Storage;
 import yoyo.exception.YoyoException;
 import yoyo.task.TaskList;
 
+/**
+ * Main class for Yoyo app.
+ */
 public class Yoyo {
     private final Storage storage;
     private final DialogHandler dialogHandler;
