+++ src/main/java/yoyo/command/CommandEvent.java
@@ -1,5 +1,7 @@
 package yoyo.command;
 
+import java.time.LocalDateTime;
+
 import yoyo.core.DialogHandler;
 import yoyo.core.Parser;
 import yoyo.core.Storage;
@@ -8,9 +10,15 @@ import yoyo.task.Event;
 import yoyo.task.Task;
 import yoyo.task.TaskList;
 
-import java.time.LocalDateTime;
-
+/**
+ * A command subclass representing "event" command.
+ */
 public class CommandEvent extends Command {
+    /**
+     * Constructor for "event" command class.
+     *
+     * @param inputTokens Array of string tokens constructed from user input.
+     */
     public CommandEvent(String[] inputTokens) {
         super(inputTokens);
     }
