+++ src/main/java/yoyo/task/Todo.java
@@ -1,5 +1,8 @@
 package yoyo.task;
 
+import static yoyo.utility.Constant.COMMA_SEPARATOR;
+import static yoyo.utility.Constant.WHITESPACE;
+
 /**
  * A subclass of Task of todo type.
  */
@@ -17,7 +20,7 @@ public class Todo extends Task {
     /**
      * Constructor for the Todo class with name and isDone parameter.
      *
-     * @param name Name of Todo.
+     * @param name   Name of Todo.
      * @param isDone Completion status of task.
      */
     public Todo(String name, boolean isDone) {
@@ -27,7 +30,9 @@ public class Todo extends Task {
     /**
      * Constructor for the Todo class with name, isDone and tags parameter.
      *
-     * @param name Name of Todo.
+     * @param name   Name of Todo.
+     * @param isDone Completion status of this task.
+     * @param tags   Array of tags belonging to this task.
      */
     public Todo(String name, boolean isDone, String[] tags) {
         super(name, isDone, tags);
@@ -50,7 +55,21 @@ public class Todo extends Task {
      */
     @Override
     public String showStatus() {
-        return super.showStatus();
+        return printType()
+                + printCompletionStatus()
+                + WHITESPACE
+                + name
+                + WHITESPACE
+                + showTags();
+    }
+
+    @Override
+    public String showStatusWrite() {
+        return this.printType()
+                + this.printCompletionStatus()
+                + COMMA_SEPARATOR
+                + this.name
+                + this.showTagsWriteFormat();
     }
