+++ src/main/java/yoyo/utility/Constant.java
@@ -1,5 +1,8 @@
 package yoyo.utility;
 
+/**
+ * Constant class that contains some commonly used constants in Yoyo.
+ */
 public class Constant {
     public static final String COMMA_SEPARATOR = ", ";
     public static final Character NEWLINE_CHAR = '\n';
